elisa2/requirements.txt
altair==6.0.0
annotated-types==0.7.0
attrs==26.1.0
blinker==1.9.0
bokeh==3.9.0
branca==0.8.2
cachetools==7.0.5
certifi==2026.2.25
cffi==2.0.0
cftime==1.6.5
charset-normalizer==3.4.6
click==8.3.1
cloudpickle==3.1.2
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.5
cycler==0.12.1
dask==2026.3.0
distributed==2026.3.0
earthengine-api==1.7.18
filelock==3.25.2
folium==0.20.0
fonttools==4.62.1
fsspec==2026.2.0
geographiclib==2.1
geopy==2.4.1
gitdb==4.0.12
GitPython==3.1.46
google-api-core==2.30.0
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-cloud-core==2.5.0
google-cloud-storage==3.10.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.73.0
httplib2==0.31.2
idna==3.11
importlib_metadata==9.0.0
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.5.0
locket==1.0.0
lz4==4.4.5
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
msgpack==1.1.2
narwhals==2.18.0
netCDF4==1.7.4
networkx==3.6.1
numpy==2.4.3
packaging==26.0
pandas==2.3.3
partd==1.4.2
pillow==12.1.1
plotly==6.6.0
proto-plus==1.27.1
protobuf==6.33.6
psutil==7.2.2
pyarrow==23.0.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pydeck==0.9.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
smmap==5.0.3
sortedcontainers==2.4.0
streamlit==1.55.0
streamlit-folium==0.26.2
sympy==1.14.0
tblib==3.2.2
tenacity==9.1.4
threadpoolctl==3.6.0
toml==0.10.2
toolz==1.1.0
torch==2.10.0
tornado==6.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
watchdog==6.0.0
xarray==2026.2.0
xyzservices==2025.11.0
zict==3.0.0
zipp==3.23.0
